[AIRFLOW-5896] Move email stuff from tests/ccore.py
